refactor: order integration projections in folders
