+++ src/BuildingRegistry.Projections.Integration/BuildingUnit/LatestItem/BuildingUnitLatestItemExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace BuildingRegistry.Projections.Integration
+﻿namespace BuildingRegistry.Projections.Integration.BuildingUnit.LatestItem
 {
     using System;
     using System.Threading;
