+++ src/BuildingRegistry.Projections.Integration/BuildingUnit/Version/BuildingUnitAddressLegacyVersion.cs
@@ -1,37 +1,39 @@
-﻿namespace BuildingRegistry.Projections.Integration
+﻿namespace BuildingRegistry.Projections.Integration.BuildingUnit.Version
 {
+    using System;
     using BuildingRegistry.Infrastructure;
     using Microsoft.EntityFrameworkCore;
     using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-    public sealed class BuildingUnitAddressVersion
+    public sealed class BuildingUnitAddressLegacyVersion
     {
         public long Position { get; set; }
-        public int BuildingUnitPersistentLocalId { get; set; }
-        public int AddressPersistentLocalId { get; set; }
 
-        public BuildingUnitAddressVersion()
+        public int BuildingUnitPersistentLocalId { get; set; }
+        public Guid AddressId { get; set; }
+
+        public BuildingUnitAddressLegacyVersion()
         { }
 
-        public BuildingUnitAddressVersion CloneAndApplyEventInfo(
+        public BuildingUnitAddressLegacyVersion CloneAndApplyEventInfo(
             long newPosition)
         {
-            var newItem = new BuildingUnitAddressVersion
+            var newItem = new BuildingUnitAddressLegacyVersion
             {
                 Position = newPosition,
                 BuildingUnitPersistentLocalId = BuildingUnitPersistentLocalId,
-                AddressPersistentLocalId = AddressPersistentLocalId
+                AddressId = AddressId
             };
 
             return newItem;
         }
     }
 
-    public sealed class BuildingUnitAddressVersionConfiguration : IEntityTypeConfiguration<BuildingUnitAddressVersion>
+    public sealed class BuildingUnitAddressLegacyVersionConfiguration : IEntityTypeConfiguration<BuildingUnitAddressLegacyVersion>
     {
-        public void Configure(EntityTypeBuilder<BuildingUnitAddressVersion> builder)
+        public void Configure(EntityTypeBuilder<BuildingUnitAddressLegacyVersion> builder)
         {
-            const string tableName = "building_unit_address_versions";
+            const string tableName = "building_unit_address_legacy_versions";
 
             builder
                 .ToTable(tableName, Schema.Integration) // to schema per type
@@ -39,18 +41,18 @@
                 {
                     x.Position,
                     x.BuildingUnitPersistentLocalId,
-                    x.AddressPersistentLocalId
+                    x.AddressId
                 });
 
             builder.Property(x => x.Position).ValueGeneratedNever();
 
             builder.Property(x => x.Position).HasColumnName("position");
             builder.Property(x => x.BuildingUnitPersistentLocalId).HasColumnName("building_unit_persistent_local_id");
-            builder.Property(x => x.AddressPersistentLocalId).HasColumnName("address_persistent_local_id");
+            builder.Property(x => x.AddressId).HasColumnName("address_id");
 
             builder.HasIndex(x => x.Position);
             builder.HasIndex(x => x.BuildingUnitPersistentLocalId);
-            builder.HasIndex(x => x.AddressPersistentLocalId);
+            builder.HasIndex(x => x.AddressId);
         }
     }
 }
