+++ src/BuildingRegistry.Projections.Integration/BuildingUnit/Version/BuildingUnitAddressVersion.cs
@@ -1,39 +1,37 @@
-﻿namespace BuildingRegistry.Projections.Integration
+﻿namespace BuildingRegistry.Projections.Integration.BuildingUnit.Version
 {
-    using System;
     using BuildingRegistry.Infrastructure;
     using Microsoft.EntityFrameworkCore;
     using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-    public sealed class BuildingUnitAddressLegacyVersion
+    public sealed class BuildingUnitAddressVersion
     {
         public long Position { get; set; }
-
         public int BuildingUnitPersistentLocalId { get; set; }
-        public Guid AddressId { get; set; }
+        public int AddressPersistentLocalId { get; set; }
 
-        public BuildingUnitAddressLegacyVersion()
+        public BuildingUnitAddressVersion()
         { }
 
-        public BuildingUnitAddressLegacyVersion CloneAndApplyEventInfo(
+        public BuildingUnitAddressVersion CloneAndApplyEventInfo(
             long newPosition)
         {
-            var newItem = new BuildingUnitAddressLegacyVersion
+            var newItem = new BuildingUnitAddressVersion
             {
                 Position = newPosition,
                 BuildingUnitPersistentLocalId = BuildingUnitPersistentLocalId,
-                AddressId = AddressId
+                AddressPersistentLocalId = AddressPersistentLocalId
             };
 
             return newItem;
         }
     }
 
-    public sealed class BuildingUnitAddressLegacyVersionConfiguration : IEntityTypeConfiguration<BuildingUnitAddressLegacyVersion>
+    public sealed class BuildingUnitAddressVersionConfiguration : IEntityTypeConfiguration<BuildingUnitAddressVersion>
     {
-        public void Configure(EntityTypeBuilder<BuildingUnitAddressLegacyVersion> builder)
+        public void Configure(EntityTypeBuilder<BuildingUnitAddressVersion> builder)
         {
-            const string tableName = "building_unit_address_legacy_versions";
+            const string tableName = "building_unit_address_versions";
 
             builder
                 .ToTable(tableName, Schema.Integration) // to schema per type
@@ -41,18 +39,18 @@
                 {
                     x.Position,
                     x.BuildingUnitPersistentLocalId,
-                    x.AddressId
+                    x.AddressPersistentLocalId
                 });
 
             builder.Property(x => x.Position).ValueGeneratedNever();
 
             builder.Property(x => x.Position).HasColumnName("position");
             builder.Property(x => x.BuildingUnitPersistentLocalId).HasColumnName("building_unit_persistent_local_id");
-            builder.Property(x => x.AddressId).HasColumnName("address_id");
+            builder.Property(x => x.AddressPersistentLocalId).HasColumnName("address_persistent_local_id");
 
             builder.HasIndex(x => x.Position);
             builder.HasIndex(x => x.BuildingUnitPersistentLocalId);
-            builder.HasIndex(x => x.AddressId);
+            builder.HasIndex(x => x.AddressPersistentLocalId);
         }
     }
 }
