+++ src/BuildingRegistry.Projections.Integration/BuildingUnit/Version/BuildingUnitVersionProjections.cs
@@ -1,4 +1,4 @@
-﻿namespace BuildingRegistry.Projections.Integration
+﻿namespace BuildingRegistry.Projections.Integration.BuildingUnit.Version
 {
     using System.Collections.ObjectModel;
     using System.Linq;
@@ -6,10 +6,10 @@
     using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
     using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
     using Be.Vlaanderen.Basisregisters.Utilities.HexByteConvertor;
-    using Building;
-    using Building.Events;
-    using Converters;
-    using Infrastructure;
+    using BuildingRegistry.Building;
+    using BuildingRegistry.Building.Events;
+    using BuildingRegistry.Projections.Integration.Converters;
+    using BuildingRegistry.Projections.Integration.Infrastructure;
     using Microsoft.Extensions.Options;
 
     [ConnectedProjectionName("Integratie gebouweenheid latest item")]
